Sem241/Software Engineering/src/backend/requirements.txt
Flask==2.2.2
Flask-Cors==3.1.1
PyPDF2==2.12.1
python-docx==0.8.11
